+++ FinallyNewLab/FinallyNewLab/Base.h
@@ -12,6 +12,6 @@ void SwapBorders(int32_t&, int32_t&);
 void RandomMatrix(int32_t**&, int32_t, int32_t, int32_t, int32_t);
 void InputMatrix(int32_t**&, int32_t, int32_t);
 void PrintMatrix(int32_t**&, int32_t, int32_t);
-void DeleteMatrix(double**&, int32_t);
+void DeleteMatrix(int32_t**&, int32_t);
 char InputChoose(char&);
 #endif
